% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-xp.R
\name{new_xp}
\alias{new_xp}
\title{Create a regexpert object}
\usage{
new_xp(x)
}
\arguments{
\item{x}{A character string representing a regex pattern.}
}
\value{
An object of class \code{xp}.
}
\description{
This is an internal constructor that converts a string into a \code{regexpert} object.
It allows for custom printing and ensures operators handle patterns correctly.
}
\keyword{internal}
