% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-xp.R
\name{print.xp}
\alias{print.xp}
\title{Print a regexpert pattern}
\usage{
\method{print}{xp}(x, ...)
}
\arguments{
\item{x}{A regexpert object.}

\item{...}{Unused.}
}
\description{
Custom print method to display \code{xp} objects cleanly in the console.
}
