% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_action_view.r
\name{xp_action_view}
\alias{xp_action_view}
\title{View the generated regex string}
\usage{
xp_action_view(pattern)
}
\arguments{
\item{pattern}{A \code{regexpert} object.}
}
\description{
Prints the current regex pattern to the console in a clean format.
}
\examples{
xp_build_digits() \%>\% xp_op_repeat(3) \%>\% xp_action_view()
}
