% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_build_alnum.r
\name{xp_build_alnum}
\alias{xp_build_alnum}
\title{Build an alphanumeric pattern}
\usage{
xp_build_alnum(pattern = NULL, n = NULL)
}
\arguments{
\item{pattern}{(Optional) An existing pattern from the pipe.}

\item{n}{Exact number of characters.}
}
\value{
A character string.
}
\description{
Build an alphanumeric pattern
}
\examples{
xp_build_alnum(n = 5)
xp_literal("#") \%>\% xp_build_alnum(n = 3)
}
