% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_build_any.r
\name{xp_build_any}
\alias{xp_build_any}
\title{Build an "any character" pattern}
\usage{
xp_build_any(pattern = NULL)
}
\arguments{
\item{pattern}{(Optional) An existing pattern from a pipe.}
}
\value{
A \code{regexpert} object.
}
\description{
Matches any single character except for a newline (the regex dot).
}
\examples{
# Match anything between two brackets
xp_build_literal("[") \%>\% xp_build_any() \%>\% xp_op_repeat(0, Inf) \%>\% xp_build_literal("]")
}
