% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_build_digits.r
\name{xp_build_digits}
\alias{xp_build_digits}
\title{Build a digit pattern}
\usage{
xp_build_digits(pattern = NULL, n = NULL)
}
\arguments{
\item{pattern}{(Optional) An existing pattern from the pipe.}

\item{n}{Exact number of digits.}
}
\value{
A character string.
}
\description{
Build a digit pattern
}
\examples{
xp_build_digits(n = 3)
xp_literal("Area Code: ") \%>\% xp_build_digits(3)
}
