% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_build_letters.r
\name{xp_build_letters}
\alias{xp_build_letters}
\title{Build a letter pattern}
\usage{
xp_build_letters(pattern = NULL, case = "both", n = NULL)
}
\arguments{
\item{pattern}{(Optional) An existing pattern from the pipe.}

\item{case}{"both" (default), "lower", or "upper".}

\item{n}{Exact number of repetitions.}
}
\value{
A character string.
}
\description{
Build a letter pattern
}
\examples{
xp_build_letters(case = "upper", n = 3)
xp_build_digits(1) \%>\% xp_build_letters(case = "lower", n = 2)
}
