% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_build_literal.r
\name{xp_build_literal}
\alias{xp_build_literal}
\title{Build or add literal text}
\usage{
xp_build_literal(pattern = NULL, literal = NULL)
}
\arguments{
\item{pattern}{(Optional) An existing pattern from a pipe.}

\item{literal}{The exact text/symbols you want to match.}
}
\value{
A \code{regexpert} object.
}
\description{
Converts raw text into a safe regex pattern by escaping special characters.
}
\examples{
# Safely match a price with a dollar sign and decimal
xp_build_literal("$") \%>\% xp_build_digits() \%>\% xp_build_literal(".") \%>\% xp_build_digits()

# Starting a chain with a literal
xp_build_literal("v1.0") \%>\% xp_wrap_boundary()
}
