% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_build_standard.r
\name{xp_build_standard}
\alias{xp_build_standard}
\title{Build a standard Pattern
Access a library of pre-defined, complex regex patterns for common data types.}
\usage{
xp_build_standard(type)
}
\arguments{
\item{type}{A character string specifying the pattern.
Options include: \code{"email"}, \code{"phone"}, \code{"zip"}, \code{"date_iso"}, \code{"date_us"}, \code{"credit_card"}, \code{"ipv4"}, \code{"hex_color"}, \code{"time_24"}, \code{"ssn"}, \code{"mac_address"}.}
}
\value{
A \code{regexpert} object.
}
\description{
Build a standard Pattern
Access a library of pre-defined, complex regex patterns for common data types.
}
\examples{
xp_build_standard("email")
xp_build_standard("zip") \%>\% xp_action_find("My code is 83440")
}
