% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_build_whitespace.r
\name{xp_build_whitespace}
\alias{xp_build_whitespace}
\title{Build a whitespace pattern}
\usage{
xp_build_whitespace(pattern = NULL)
}
\arguments{
\item{pattern}{(Optional) An existing pattern from a pipe.}
}
\value{
A \code{regexpert} object.
}
\description{
Matches spaces, tabs, and line breaks.
}
\examples{
# Match two words separated by whitespace
xp_build_letters() \%>\% xp_build_whitespace() \%>\% xp_build_letters()
}
