% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_literal.r
\name{xp_literal}
\alias{xp_literal}
\title{Add literal text to a pattern}
\usage{
xp_literal(pattern = NULL, literal = NULL)
}
\arguments{
\item{pattern}{(Optional) An existing regex string from the pipe.}

\item{literal}{The exact text/symbols you want to add.}
}
\value{
A character string.
}
\description{
Add literal text to a pattern
}
