% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_op_either.r
\name{xp_op_either}
\alias{xp_op_either}
\title{Add an alternative pattern (OR)}
\usage{
xp_op_either(pattern, alternative)
}
\arguments{
\item{pattern}{A \code{regexpert} object or character string from the pipe.}

\item{alternative}{The alternative pattern or string to match.}
}
\value{
A \code{regexpert} object.
}
\description{
Adds an "either/or" condition to the pattern. This is equivalent to the
regex pipe \code{|} operator.
}
\examples{
# Match "cat" or "dog"
xp_build_literal("cat") \%>\% xp_op_either("dog")

# Match a digit or a letter
xp_build_digits() \%>\% xp_op_either(xp_build_letters())
}
