% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_op_optional.r
\name{xp_op_optional}
\alias{xp_op_optional}
\title{Make a pattern optional}
\usage{
xp_op_optional(pattern)
}
\arguments{
\item{pattern}{A character string.}
}
\value{
A character string.
}
\description{
Tells the engine that the preceding pattern may appear 0 or 1 times.
}
\examples{
xp_build_digits(3) \%>\% xp_op_optional()
}
