% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_build_standard.r
\name{xp_register}
\alias{xp_register}
\title{Register a Custom Standard}
\usage{
xp_register(name, pattern)
}
\arguments{
\item{name}{A character string to name the pattern.}

\item{pattern}{A regex string or regexpert object.}
}
\description{
Adds a custom regex pattern to the global library for the current session.
}
