% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_view.r
\name{xp_view}
\alias{xp_view}
\title{View regex matches}
\usage{
xp_view(text, pattern)
}
\arguments{
\item{text}{The character vector to search.}

\item{pattern}{The regex pattern to test.}
}
\value{
A list of matches found.
}
\description{
Displays which parts of the text are being captured by your pattern.
}
\examples{
"The cat in the hat" \%>\% xp_view(xp_build_letters(case = "lower"))
}
