% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_wrap_end.r
\name{xp_wrap_end}
\alias{xp_wrap_end}
\title{Anchor pattern to the end}
\usage{
xp_wrap_end(pattern)
}
\arguments{
\item{pattern}{A character string.}
}
\value{
A character string with a '$' suffix.
}
\description{
Forces the pattern to match only if it appears at the end of the string.
}
\examples{
xp_build_letters() \%>\% xp_wrap_end()
}
