% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_wrap_start.r
\name{xp_wrap_start}
\alias{xp_wrap_start}
\title{Anchor pattern to the start}
\usage{
xp_wrap_start(pattern)
}
\arguments{
\item{pattern}{A character string.}
}
\value{
A character string with a '^' prefix.
}
\description{
Forces the pattern to match only if it appears at the beginning of the string.
}
\examples{
xp_build_digits(3) \%>\% xp_wrap_start()
}
