% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_build_alnum.r
\name{xp_build_alnum}
\alias{xp_build_alnum}
\title{Build an alphanumeric pattern}
\usage{
xp_build_alnum(pattern = NULL)
}
\arguments{
\item{pattern}{(Optional) An existing pattern from a pipe.}
}
\value{
A \code{regexpert} object.
}
\description{
Matches any single letter (A-Z, a-z) or digit (0-9).
}
\examples{
# Match a string of 8 alphanumeric characters (like a serial number)
xp_build_alnum() \%>\% xp_op_repeat(8)
}
