% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_build_any.r
\name{xp_build_any}
\alias{xp_build_any}
\title{Build a wildcard pattern}
\usage{
xp_build_any(pattern = NULL, n = NULL)
}
\arguments{
\item{pattern}{(Optional) An existing pattern from the pipe.}

\item{n}{Exact number of characters.}
}
\value{
A character string.
}
\description{
Build a wildcard pattern
}
\examples{
xp_literal("A") \%>\% xp_build_any() \%>\% xp_literal("C")
xp_build_any(n = 10)
}
