% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_build_whitespace.r
\name{xp_build_whitespace}
\alias{xp_build_whitespace}
\title{Build a whitespace pattern}
\usage{
xp_build_whitespace(pattern = NULL, n = NULL)
}
\arguments{
\item{pattern}{(Optional) An existing pattern from the pipe.}

\item{n}{Exact number of spaces/tabs.}
}
\value{
A character string.
}
\description{
Build a whitespace pattern
}
\examples{
xp_literal("Word") \%>\% xp_build_whitespace() \%>\% xp_literal("Word")
xp_build_whitespace(n = 2)
}
