% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_action_find.r
\name{xp_find}
\alias{xp_find}
\title{Find matches in a string}
\usage{
xp_find(pattern, string)
}
\arguments{
\item{pattern}{A \code{regexpert} object or character string.}

\item{string}{A character vector to search within.}
}
\value{
A character vector of matches.
}
\description{
Searches a character vector for matches to the built regex pattern.
}
\examples{
my_text <- "The year is 2026."
xp_build_digits() \%>\% xp_op_repeat(4) \%>\% xp_find(my_text)
}
