% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_op_either.r
\name{xp_op_either}
\alias{xp_op_either}
\title{Match one of two patterns}
\usage{
xp_op_either(pattern_a, pattern_b)
}
\arguments{
\item{pattern_a}{The first pattern.}

\item{pattern_b}{The second pattern.}
}
\value{
A character string.
}
\description{
Provides an 'OR' logic between two different patterns.
}
\examples{
xp_op_either(xp_build_digits(3), xp_build_letters(2))
}
