% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_op_optional.r
\name{xp_op_optional}
\alias{xp_op_optional}
\title{Make a pattern optional}
\usage{
xp_op_optional(pattern)
}
\arguments{
\item{pattern}{A \code{regexpert} object or character string from the pipe.}
}
\value{
A \code{regexpert} object.
}
\description{
Specifies that the preceding pattern should occur zero or one time.
Equivalent to the \code{?} quantifier.
}
\examples{
# Match "color" or "colour"
xp_build_literal("colo") \%>\% 
  xp_build_literal("u") \%>\% 
  xp_op_optional() \%>\% 
  xp_build_literal("r")

# Optional area code
xp_build_literal("(") \%>\% xp_build_digits() \%>\% xp_op_repeat(3) \%>\% 
  xp_build_literal(") ") \%>\% 
  xp_op_optional() \%>\% 
  xp_build_digits() \%>\% xp_op_repeat(7)
}
