% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_wrap_boundary.r
\name{xp_wrap_boundary}
\alias{xp_wrap_boundary}
\title{Wrap in word boundaries}
\usage{
xp_wrap_boundary(pattern)
}
\arguments{
\item{pattern}{A \code{regexpert} object or character string from the pipe.}
}
\value{
A \code{regexpert} object.
}
\description{
Adds \code{\\b} to both sides of the pattern. This ensures that the
pattern is matched as a whole word (not inside another word).
}
\examples{
# Match "cat" but not "category" or "bobcat"
xp_build_literal("cat") \%>\% xp_wrap_boundary()
}
